+++ azure/CHANGELOG.md
@@ -6,4 +6,6 @@ This file documents changes to the `workbench-azure` library, including notes on
 
 ### Added
 
-SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.1-91d3580"`
+- Added AzureStorageService and AzureStorageManualTest
+
+SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.1-8cef579"`